fixing flows in pull secrets
