+++ frontend/public/components/secrets/create-secret/utils.ts
@@ -49,7 +49,7 @@ export const determineSecretType = (stringData): SecretType => {
   return SecretType.opaque;
 };
 
-export const getImageSecretKey = (secretType: SecretType): string => {
+export const getPullSecretFileName = (secretType: SecretType): string => {
   switch (secretType) {
     case SecretType.dockercfg:
       return '.dockercfg';
